+++ stack.h
@@ -12,21 +12,22 @@
 #define STACK_H
 
 #include <stdbool.h>
+#include <stdint.h>
 
 extern unsigned short int _stack_pointer;
-extern int _stack[];
+extern uint16_t _stack [];
 
 // Checks weather the stack if full
-bool stack_full();
+bool stack_full ( );
 
 // Pushes an item on top of the stack
-bool stack_push (int value);
+bool stack_push ( int value );
 
 // Checks weather the stack is empty
-bool stack_empty (void);
+bool stack_empty ( void );
 
 // Pulls an item out of the stack
-int stack_pop (void);
+uint16_t stack_pop ( void );
 
 
 #endif // STACK_H
